[FIX] sign: add template_id as query parameter
- Sign > Templates > Click on any of the kanban record
- Reload the view (either reload the browser, or activate
the debug, or change to dark mode on the user menu).

Before this commit, on reloading the sign template page user was
redirected back to kanban view. This occurs because, the client
action 'sign.Template' required template id (found in the context)
that was lost when reloading.

Now, template id is put in the query string of the URL as id,
in that way, when reloading, the client action 'sign.Document'
will have the needed template id. Note that, this is also
the behavior of the base_import 'ImportAction' action [1].

task-4391729

[1] : https://github.com/odoo/odoo/commit/a6801ce4aa65cd023e2a762ab5c42755a4599c95

closes odoo/enterprise#81640

X-original-commit: c735b475de4eea18bc51a56d1a8d4a93db7fb0b8
Signed-off-by: Arnaud Joset (arj) <arj@odoo.com>
Signed-off-by: Saurabh Choraria (sauc) <sauc@odoo.com>