[FIX] hr_contract_salary: exclude studio fields from compute dependencies
Add a boolean field on hr.contract via studio. Then remove that field from the model.
That last operation crashes because _compute_final_yearly_costs has a dynamic depends on it which take many
fields of the model.

While we don't want to exclude every custom fields (starting with "x_") to continue to allow sufficient extension points
we do exclude studio fields specifically, because those are not meant to be used in a complex compute method.

After this commit, there is no crash

opw-4503855

closes odoo/enterprise#80323

X-original-commit: 51f4ac6c4dad97de3e0113f64f9deaff71bde972
Signed-off-by: Bastien Fafchamps (bafa) <bafa@odoo.com>
Signed-off-by: Lucas Perais (lpe) <lpe@odoo.com>