[IMP] l10n_co_dian: Commercial Events
According to DIAN legal requirement: Electronic documents must have
several associated events that indicate the validation status of the
DIAN and the commercial status with the client.

After the issuance of the electronic Invoice, once the client receives
it, it is necessary to generate several additional events for the
invoice which notify the reception of the document, reception of goods,
acceptance, claim, etc.

This task will enable Odoo users to report commercial events of
Vendor Bills received, receive the updates for their own
Customer Invoices and make them able to Accept by Issuer if needed.

closes odoo/enterprise#83412

Task: 4242575
Signed-off-by: Laurent Smet (las) <las@odoo.com>