[IMP] l10n_in_reports: prevent creating duplicate gst return periods
Steps to Reproduce:
===================
The user creates a GST return period for example January 2025 between
1st - 10th April.
After 10th April, the user creates another GST return period for January 2025.
The system allows the creation of a duplicate record.

Reason:
=======
As per GST filing rules, the return period for the previous month can be filed
on the government portal until the 10th of the following month.
The system automatically sets the month and quarter values to the previous month
and previous quarter respectively  if the current date is between the 1st and
10th of a month, when the periodicity is set to monthly or quarterly.

Cause:
======
Both records are considered different by the system based on the quarter value
or month value and the system does not block the creation of duplicate GST
return periods.

Technical:
==========
Modified the SQL constrains into two separate constrains for checking unique
return period monthly and quarterly.
Also, changed the default method to the compute method and added quarter logic
in the creation of the Return period from the 'fetch from GSTN` button.

Changed the quarter's key because with the start month, it's considered as IFF

Task-4534812

closes odoo/enterprise#79220

Related: odoo/upgrade#7270
Signed-off-by: Josse Colpaert (jco) <jco@odoo.com>