[FIX] *_spreadsheet_*: chart panel in error during loading
When we change the chart type/change the chart doamin from the side
panel, there would be an error message `the model (odoo.model) of this
chart is not valid`. Which was wrong, the model was valid but the
data source was simply not loaded yet.

This commit instead shows a loading spinner when the data source is
loading.

closes odoo/enterprise#80175

Task: 4564072
X-original-commit: 5229c2cdc2bcf071891c775764aad36dc7d7b185
Signed-off-by: Lucas Lefèvre (lul) <lul@odoo.com>
Signed-off-by: Adrien Minne (adrm) <adrm@odoo.com>