[FIX] sale_subscription: avoid concurrent update
Scenario
========

the subscription cron that automate the payment, start a payment
transaction get the response and post process the the payment.
In the meanwhile, the webhook receive also the response and try to post
a message on the invoice in _log_message_on_linked_documents

This lead to concurrent update on the invoice

Solution
========

use _get_invoices_to_notify to remove the invoice currenctly handled by
the invoice cron from the payment notification.

The notification won't appear on the invoice in this flow, that's not a
problem. Payment notification is also set on the sale order

closes odoo/enterprise#79298

X-original-commit: 317f9cb88be27b91adc5d53eae6bf07e0db7f09d
Related: odoo/odoo#197693
Signed-off-by: Arnaud Joset (arj) <arj@odoo.com>