[IMP] approvals_purchase: log a message for RFQs creation and state
changes.

When an approval request of category `Create RFQ`
is approved, the approver can create an RFQ. This
will  create a new RFQ with the requested products
or will add the requested products to existing RFQs
that have the same vendors as the products. However, the user
who created the approval request doesn't get any
feedback on the RFQs.

To solve this issue, A message is logged to the
chatter of the approval request. The message will
show each product and the reference (name) of the
corresponding RFQ to which the product has been added.

Also, when the state of the RFQ changes, a message will be logged
to the chatter of the approval request. The message will show
the products from the approval request that are present in
the RFQ, the RFQ reference, the old state and new state of
the RFQ.

task-3941803

closes odoo/enterprise#71294

Signed-off-by: Yannick Tivisse (yti) <yti@odoo.com>