[FIX] web_grid: take into account the action context to update grid cell
Before this commit, the context of the action was just ignored in the
grid view and so the default values stored inside that context are not
taken into account during `grid_update_cell` method.

Steps to reproduce the issue
----------------------------
0. Install timesheet_grid module
1. Open Employees app
2. Click on an employee to open the form view
3. Click on the timesheets smart button to view the timesheet grid view
   for that employee
4. Navigate to a week that already contains hours for the current user
5. Update the grid cell containing a timesheet to reduce the number of
   hours for instance
6. Update another cell without any timesheets linked and set a hour

Current Behavior
----------------
After refreshing the page and going to the same week, the changes made
in step 5 and 6 are no longer applied. The reason is because those
changes have generated new timesheets for the current user instead of
the employee displayed in the form view before doing the step 3.

Expected behavior
-----------------
The grid view should be unchanged after refreshing the page, the changes
made in 5 and 6 should be applied without any issues.

opw-4491850

closes odoo/enterprise#79300

X-original-commit: e9885a29de14a0a94fd253b8d609b28224e786a8
Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>