[IMP] {test_}whatsapp: avoid delayed message sending in tests
Whatsapp uses `_trigger` a lot in the sending process.
This offloads the actual sending to a cron worker
instead of blocking the request thread.

During tests we do not care about blocking the thread.
It is rarely interesting to test the behavior of the crons being triggered.

Every call to IrCron._trigger is replaced with IrCron.method_direct_trigger
during tests. This promotes testing for the actual end state of each message
in tests, rather than most messages being "outgoing".

We also prevent auto-commiting during tests
to avoid breaking test cleanup.

task-3547088

X-original-commit: 3d25d493bbc6e256844b0f02a640e72f102e9195
Part-of: odoo/enterprise#85176
Signed-off-by: Thibault Delavallee (tde) <tde@openerp.com>
Signed-off-by: Renaud Thiry (reth) <reth@odoo.com>