[FIX] website_sale_subscription : Show pricing table only when having more than one price
Steps to reproduce:
	- Create a recurring product with only one recurring plan
	- Go to the product's page in the website
	- Notice the pricing table is shown with only one plan

Cause:
In the pricing view in the website we are showing the pricing table
just after checking if we have pricings or not.

Fix:
We are now checking if the pricings we have is more than one we will
show the table otherwise will not show it.

opw-4559071

closes odoo/enterprise#79414

X-original-commit: 34ee3c9540de6744a817487261533bb087769789
Signed-off-by: Arnaud Joset (arj) <arj@odoo.com>
Signed-off-by: Youssef Bashandy (yoba) <yoba@odoo.com>