[REF] point_of_sale,*: prefer precision (rounding) over decimals
Numbers in the app are normally compared considering the number of decimals. We
argue in this commit that precision (rounding) value is a better parameter to
use for comparisons (and rounding) because it can support arbitrary values such
as 0.05, 0.1, 5, 10, etc (and not just powers of 10).

So to homogenize the codebase, we replace all number comparisons to use
precision (rounding) value.

The comparison methods are introduced to client models that are typically used
to compare numbers such as:

- res.currency -- used to compare prices (amounts)
- uom.uom -- used to compare quantities
- decimal.precision -- depends on what it's defined to compare

Both res.currency and uom.uom have a `rounding` attribute which are used by
default in the `Numbers` base class. The decimal.precision has `digits` which
will be used to compute `rounding` value.

closes odoo/enterprise#79167

Related: odoo/odoo#197447
Signed-off-by: Vlad Stroia (vlst) <vlst@odoo.com>