[FIX] project_enterprise: avoid error when opening Gantt view with empty working schedule
**Issue**
When setting the working schedule to fully flexible, opening the Gantt view is not possible as an error is displayed

Steps to Reproduce:
Login as Mitchell Admin
Install Employees and Project apps
Go to Employees > Contracts
Select a contract from Mitchell Admin
Empty the Working Schedule Field
Set Planning or Attendances as work entries, then save
Go to project
Select a project
Switch to the Gantt view
An error message appears
Expected Behavior: Gantt view should be displayed without errors.
Actual Behavior: Gantt view is not displayed, and an error message appears.

**Root Cause**

The issue occurs because before processing the work intervals, the system checks for the presence and uniqueness of a calendar. If a user has no assigned schedule, the function attempts to process a None value, which results in an error when calling ensure_one in _attendance_intervals_batch.

**Fix**
To prevent the error, users without a working schedule are skipped when computing valid work intervals. If a user has no assigned schedule, they will not be included in _work_intervals_batch, ensuring the fu

opw-4462241

closes odoo/enterprise#82909

X-original-commit: 08c3e6655399926864d63c2150cfde88a5cb34fe
Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>