[IMP] sign: remove 'available' field in options
This commit removes the 'available' field from the 'sign.item.option' model
as it is not used anymore neither its logic. After odoo/enterprise#75291
we create options on demand and reuse the previous options for avoiding
duplicating it, with the available field being True or False.

issue-from: 75291
task-4387560

closes odoo/enterprise#79161

Related: odoo/upgrade#7212
Signed-off-by: Arnaud Joset (arj) <arj@odoo.com>