[FIX] documents: fix documents search panel and record container
Fix documents search panel onWillStart
======================================

To prevent client-side errors when landing on the Documents' page, we
need to wait for the result of the (async) calls to the searchModel made
from the DocumentsSearchPanel.

Re-render DocumentsDetailPanel on Container record load
=======================================================

For some use cases, we need the re-render the detail panel, when the
Container record load, to get the panel fields properly updated.

task-4512349

Part-of: odoo/enterprise#77890
Related: odoo/upgrade#7133
Signed-off-by: Florian Charlier (flch) <flch@odoo.com>