[IMP] appointment: keep display consistency on validation page
Purpose
=======
Keep a display consistency for the user/resource on the validation page
and show the pictures when multiple resources are booked.

Specification
=============
When the appointment type have the "resource_manage_capacity" set to True,
it is the only configuration where multiple resources can be booked.
In this case, on the validation page, display the resources in the
form details no matter if there is one or multiple resources booked
to keep a display consistency for the appointment type.
If the avatars_display is in 'show picture', display the pictures next to
the resource names.

When the "resource_manage_capacity" is set to False (when based on users or
based on resources but we can only book one at a time), display the user or resource
on the right on the validation page (with the description and/or image if in "Show picture").
Only exception is when there is no description and the appointment type configuration
is in "No Picture", in this case to prevent displaying the user/resource name alone
on the right, displaying the user/resource in the form details.

When the config is "auto-assign" and "no picture":
- if based on resources, completely hide it from the validation page.
i.e. if a restaurant wants to auto-assign tables, no need for the user to know which one.
- if based on users, hide its details on the right of the validation page but keep them
in the form details section.
i.e. if we book an appointment with a Sale, it's important to know with who but no need to
have the floating info without pictures on the right as they already are in the form details section.

Task-4395325

closes odoo/enterprise#75492

Signed-off-by: Jérémy Hennecart (jeh) <jeh@odoo.com>