[FIX] industry_fsm_sale: don't copy under warranty field when duplicating task
Before this commit, the under warranty field is always copied when the
user duplicates a task which it does not make sense since the new task
could not have exactly the same context and the intervention could be
different.

This commit avoid copying the under warranty field when the user
duplicates a task to let the user decide if the new task should be also
in warranty or not.

opw-4571460

closes odoo/enterprise#79332

X-original-commit: 642264a6db120ce868e8b0ae9d1c53ef474c1136
Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>