[FIX] stock_barcode_mrp: add components with different unit category on MO
Steps to reproduce:

- Create 3 products: Final product (FP), components COMP
  FP using the UOM category of Unit and COMP using Weigth for instance.
- Create and confirm an MO for 1 unit of FP
- Process the MO from the barcode app.
- Click on "Add Component" select COMP as a product and set 1 unit
- Confirm the additional move line
> Invalid operation: "The unit of measure Unit defined on the order line
doesn't belong to the same category as the unit of measure g defined on
the product. Please correct the unit of measure defined on the order
line or on the product. They should belong to the same category."

Cause of the issue:

When processing an MO from the barcode app and adding a component the
method `_getNewLineDefaultContext` will be called to set the default
context of that new line before creation:
https://github.com/odoo/enterprise/blob/83268aa3a5a7afdc8c09b0ac9bbee2b673fd25e9/stock_barcode_mrp/static/src/models/barcode_mrp_model.js#L367-L372
However, in this flow, `this.record` is an `mrp.production` and its
`product_uom_id` is the product_uom_id of FP. In particular, the
`default_product_uom_id` key will be used to alter the vals_list used in
the create of our stock move line and an error will be raised since the
unit category of COMP2 is Weight and is not compatible with the default
product uom set by the context.

opw-4527706

closes odoo/enterprise#81934

X-original-commit: ead04551b293706ac2f777dedce48f9939cb364e
Signed-off-by: Steve Van Essche <svs@odoo.com>
Signed-off-by: Lancelot Semal (lase) <lase@odoo.com>