[IMP] account_report: filter invoice fully inside report period
Steps to reproduce:
1. Create a bill with date 2023-01-01
2. Create a line with start date = 2023-01-01, end date = 2023-12-31
3. Create a line with start date = 2023-01-01, end date = 2023-01-31

In the deferred report, audit the total. Both lines are shown while
only the first one should be shown.

In the same way that we do not consider invoices/bills that
are fully inside a period (i.e. move date, deferred start
and end dates all between the bounds of the report dates),
we now do it for the auditing as well.

task-4568392

X-original-commit: cb37342d70ed4acaaa78d6166a24bfea2e7208fe
Part-of: odoo/enterprise#79940
Signed-off-by: William André (wan) <wan@odoo.com>
Signed-off-by: Ricardo Gomes Rodrigues (rigr) <rigr@odoo.com>