[IMP] hr_appraisal_survey: set data survey languages
We have added a field on the model survey (lang_ids) that allow the user to
define the language in which the survey is translated. The default value for
this field is the current user language. To not limit all the survey created in
demo/data to English, we set lang_ids to False for them which means that they
are available in all installed language.

Task-3884012

closes odoo/enterprise#79268

Related: odoo/odoo#165760
Signed-off-by: Florian Charlier (flch) <flch@odoo.com>