[FIX] planning: gantt: date_start, date_end in url
The use of the parameters date_start and date_end in the url is not always
correct in the planning gantt model. For instance, if date_start and date_end
are given and are in the same week, the range selected has one day more than
expected and does not start on the first day of the week.
Here we fix/improve the usage of those parameters by removing the extra day (if
any) and localizing appropriately the ranges.

Task ID: 4500939

closes odoo/enterprise#79585

X-original-commit: fed9cb64c5560fc00c099a0926d31d1556fb64ad
Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>
Signed-off-by: Mathieu Duckerts-Antoine (dam) <dam@odoo.com>