[IMP] appointment: try to re-use similar appointment.invite records
When opening the appointment invite modal, we now attempt to find a
"identical configuration".

We only account for simple use cases, aka not multi-appointments,
not user/resource specific.

If we find an identical configuration, then the "Copy Link & Close" button of
the modal will just copy the existing record's URL instead of creating a new
record.

When the user changes anything to the configuration, we re-generate a new
random short code and will create a new record.

Task-4280698

Part-of: odoo/enterprise#79295
Signed-off-by: Warnon Aurélien (awa) <awa@odoo.com>
Co-authored-by: Aditya Sharma <adsh@odoo.com>