[FIX] l10n_be_hr_payroll_group_s: raise exception if no group_s value in company
Currently, a traceback occurs when the user tries to create a
Export to Group S report.

To reproduce this issue:
1) Install `l10n_be_hr_payroll_group_s` without demo data
2) Shift to the Belgium company
3) Create a running contract from payroll with stating date should be in past
4) Now create a Work Entry from payroll with the From date also be in the past
5) Now create a new `Export to Group S report` report from payroll/reporting
6) Make sure the date and moth are also in the past and click the `Populate`
7) Now click the `Generate Export File`

Error:- "TypeError: can only concatenate str (not "bool") to str"

The error is occurring because there is no `group_s_code` value by default
in the company. The user has to give the value from the settings.

If there is no `group_s_code` in the company, It leads to the above exception
when concatenation is done between False and a string.
Also, a userError wasalready  raised from the very next life.

https://github.com/odoo/enterprise/blob/eb1eff85d2119a34997067fedb602e963a5d6f02/l10n_be_hr_payroll_group_s/models/hr_payroll_export_group_s.py#L41-L43

So by making an extra check of the presence of the `group_s_code` in the
company. We can resolve this issue and also make the code more robust.

sentry-6289053708

closes odoo/enterprise#82235

X-original-commit: 49b3e1394318b339b12b72060315178bebfa07db
Signed-off-by: Bertrand Dossogne (bedo) <bedo@odoo.com>