[FIX] sign: fix traceback when refusing a document
Steps to reproduce:
1. Upload a document and click the "Sign Now" button.
2. Attempt to refuse the document → traceback occurs.

Issue:
A traceback error occurs.

Cause:
the functionality to cancel a document from an email was merged
into the `/sign/refuse/` controller.  when `signRequestToken` is assigned in
`sign_request_action.js` and the document is refused, the controller calls the
`cancel()` method, which updates the document's access token.  as a result, the
thankyoudialog encounters an error due to an invalid access token of a sign
request during the RPC call.

Fix:
Avoid returning http.request.not_found() which will translate into a
string and it will crash because map is not available on strings: result.map((doc) => { ... }
Make sure _refuse call cancel method which will reset the access token.

task-4551799

closes odoo/enterprise#80612

X-original-commit: 61a962ed4839096ed4f1254be3e7075d22682fd5
Signed-off-by: Arnaud Joset (arj) <arj@odoo.com>
Signed-off-by: Gurupreet Singh (gusi) <gusi@odoo.com>