[IMP] appointment: improve appointment.invite form interface
This commit adds an interface field to dissociate selecting users and
resources with different radio fields.

Indeed, it did not make sense to have a "Me" option when creating an invite for
an appointment on resources.

To keep things simple, it's just a non-stored field with an inverse that sets
the initial field value, that way the code logic stays the same.

In addition, we hide the website field with the same condition as the
appointment_type_ids field, as it's not used in that case.

Task-4280698

closes odoo/enterprise#79295

Signed-off-by: Warnon Aurélien (awa) <awa@odoo.com>
Co-authored-by: Aditya Sharma <adsh@odoo.com>