[FIX] planning: return no unavailabilities for open shifts
Steps to reproduce:
-------------------

- Install `Planning` module
- Ensure that the company has a working schedule
- Go to Planning

Issue:
------

For the `Open Shifts` resource, Saturday and Sunday are considered as
unavailable (greyed).

Cause:
------

When computing the unavailabilities for the `Open Shifts` resource, the
`gantt_unavailability` method is returning the unavailabilities for the
open shift (row with no `resID`) based on the working schedule of the
company.

Solution:
---------

Return no unavailabilities for `Open Shifts`.

task-4353378

closes odoo/enterprise#74783

Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>