[IMP] l10n_be_hr_payroll: manage salary advances
This commit improves how salary advances are managed.
Now, there is a new structure type within the CP200 structure that allows
to create a payslip with a salary advance amount as an other input.
When generating the CP200 monthly payslip, the total salary advance already
granted is added as an editable other input 'Salary Advance Recovery',
which is deducted from the total net.

This commit also renamed the existing 'Salary Advance' input type and
rule (which represented a recovery) to 'Salary Advance Recovery' to
avoid confusion with the initial salary advance.

task-3614756

closes odoo/enterprise#79380

Signed-off-by: Yannick Tivisse (yti) <yti@odoo.com>