[FIX] documents: fix drag and hide search panel
Issue 1, Drag and Drop
Reproduce:
1. Try drag & drop in search panel.
2. Either it doesn't work, or it crashes.

Issue 2,
Reproduce:
1. Go to mobile view
2. Select a document
3. `Folders`, the search panel, is still visible

In 0d904521, the `root` property was redefined in document's
search panel and overrode the `root` `ref` from the search
panel template.
Furthermore, if we renamed the property, if at some point we
are without selection while in "All", and then open a folder
and select a document, the side panel won't be hidden either.

This commit then reverts these changes.

We also take the opportunity, while in master, to clean some
unrelated dead code.

closes odoo/enterprise#79428

Signed-off-by: Stéphane Debauche (std) <std@odoo.com>