[FIX] documents: fix flicker when unselecting records
This commit fixes the visual glitch appearing when unselecting records in
Kanban and List views. The flicker was caused by the SearchBar appearing
in the ControlPanel when DocumentsAction component was still mounted and
visible for a fraction of seconds (because computing actions related to
the current selection might take some time).

Now, by adding the t-if on the root node, as since as there is no selection,
the element is hidden from the view.

task-4582340

X-original-commit: 30c1f4b70b78a9da3eacc33d719703a46fb53566
Part-of: odoo/enterprise#79752
Related: odoo/odoo#198422
Signed-off-by: Florian Charlier (flch) <flch@odoo.com>
Signed-off-by: Luca Vitali (luvi) <luvi@odoo.com>