[FIX] documents: fix crash when selecting a range of records
This commits fixes the way a single card selection is handled in Kanban.
Since we have to keep the lastCheckedRecord in the Kanban renderer, we
must use the toggleSelection props given, instead of using directly the
toggleSelection method on the record (which currently can't register this
info).

DocumentsKanbanRecord has been improved, by adding the record.id on every
card, instead of only adding it on draggable views.

A QUnit test has been introduced, for Kanban selection testing.

task-4582340

X-original-commit: 653ff8b51b401bea7f71eb12f035d0058d1d29ec
Part-of: odoo/enterprise#79752
Related: odoo/odoo#198422
Signed-off-by: Florian Charlier (flch) <flch@odoo.com>
Signed-off-by: Luca Vitali (luvi) <luvi@odoo.com>