[IMP] account: bills duplicate detection
Purpose
=======

Recent changes have made bill encoding easier than ever
(most notably, a bill date is set by default if left empty),
but offering a robust duplicate bill detection is more important
than ever in the modern electronic invoicing world, where bills
are more and more handled automatically.

We want to make sure that:
- Potential duplicates remain visible at all time, including in list view
- Automated actions like auto posting are not done when a potential duplicate is detected.

Specs
=====

1.  Do not hide the warning on potential bill duplicate for posted invoices.
The information remains pertinent.

2. When a move should be posted automatically (by the Auto-post feature on
account.move or by the OCR auto-post), but a potential duplicate is found,
prevent the auto-post from happening. Log the information in the chatter:
"Auto-post was disabled on this invoice because a potential duplicate was detected."

3. In list view put the reference in red background when a
potential duplicate on that invoice is detected.

task-4578751

closes odoo/enterprise#81094

X-original-commit: d46140e994e5a774aba43e053c19425f1d07d6a4
Related: odoo/odoo#201071
Signed-off-by: Ruben Gomes (rugo) <rugo@odoo.com>
Signed-off-by: Ricardo Gomes Rodrigues (rigr) <rigr@odoo.com>