[IMP] web_gantt, *: smaller pills, regular row height
This commit adjusts the height of pills and rows for a more uniform and
compact display. The pill height is reduced from 35px to 25px, allowing
more information to fit on the screen. Row height is now standardized to
25px per pill (or 25px when empty), with an additional 10px of empty
space. This ensures consistent row heights, whether a row contains pills
or not.

Part of task-4571228

X-original-commit: a641be38cbc3f460d0c4c299b6e9625477818899
Part-of: odoo/enterprise#80173
Signed-off-by: Mathieu Duckerts-Antoine (dam) <dam@odoo.com>