[FIX] account_followup: fix extra spaces in chatter and pdf
In the follow up process, while the wizard's text is displayed correctly, the text in the chatter and in the pdf would contain too many carriage returns.

The error is linked to the use of the text widget in the xml, which doesn't seem necessary in this case and is therefore removed.

task-4574941

closes odoo/enterprise#80153

X-original-commit: de745d1a6964bfdf233178041ea9c607abf604e1
Signed-off-by: Ruben Gomes (rugo) <rugo@odoo.com>
Signed-off-by: Camille Spiritus (casp) <casp@odoo.com>